requirements.txt
pandas>=1.3.0
numpy>=1.20.0
ccxt>=3.1.0
python-dotenv>=0.19.0
PyQt5>=5.15.0
matplotlib>=3.4.0
flask>=2.0.0
flask-cors>=3.0.0
flask-login>=0.6.2
werkzeug>=2.0.0
